REFATOR :: PostMaterialInput 파일분할
